+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fix timeout error during `setup wifi` on Cores. ([#144](https://github.com/spark/particle-cli/issues/144))
 * Standardize Wi-Fi question order. ([#19](https://github.com/spark/particle-cli/issues/19))
 * Fix `list` if no devices are found.
+* Generate keys with `keys new` if no DFU device found but `--protocol` specified.
 
 ## 1.9.3 - 20 Jan 2016